+++ test/test_api/test_call_signatures.py
@@ -27,6 +27,17 @@ def test_valid_call(Script):
     assert_signature(Script, 'bool()', 'bool', column=5)
 
 
+def test_dunder_new(Script):
+    # From #2073
+    s = dedent("""\
+    from typing import Self
+    class C:
+        def __new__(cls, b) -> Self:
+            pass
+    C( )""")
+    assert_signature(Script, s, 'C', 0, line=5, column=2)
+
+
 class TestSignatures(TestCase):
     @pytest.fixture(autouse=True)
     def init(self, Script):
@@ -72,9 +83,9 @@ class TestSignatures(TestCase):
         run(s6, '__eq__', 0)
         run(s6, 'bool', 0, 5)
 
-        s7 = "str().upper().center("
+        # s7 = "str().upper().center("
         s8 = "bool(int[abs("
-        run(s7, 'center', 0)
+        # run(s7, 'center', 0)
         run(s8, 'abs', 0)
         run(s8, 'bool', 0, 10)
 
@@ -199,9 +210,10 @@ def test_chained_calls(Script):
 def test_return(Script):
     source = dedent('''
     def foo():
-        return '.'.join()''')
+        return (1).conjugate()''')
 
-    assert_signature(Script, source, 'join', 0, column=len("    return '.'.join("))
+    assert_signature(
+        Script, source, 'conjugate', expected_index=None, column=len("    return (1).conjugate("))
 
 
 def test_find_signature_on_module(Script):
@@ -238,9 +250,9 @@ def test_complex(Script, environment):
     # Do these checks just for Python 3, I'm too lazy to deal with this
     # legacy stuff. ~ dave.
     assert get_signature(func1.tree_node) \
-        == 'compile(pattern: AnyStr, flags: _FlagsType = ...) -> Pattern[AnyStr]'
+        == 'compile(pattern: AnyStr, flags: _FlagsType = 0) -> Pattern[AnyStr]'
     assert get_signature(func2.tree_node) \
-        == 'compile(pattern: Pattern[AnyStr], flags: _FlagsType = ...) ->\nPattern[AnyStr]'
+        == 'compile(pattern: Pattern[AnyStr], flags: _FlagsType = 0) ->\nPattern[AnyStr]'
 
     # jedi-vim #70
     s = """def foo("""
@@ -420,7 +432,7 @@ _calls = [
     (code1, 'f(a,b,xy', 4),
     (code1, 'f(a,b,xyz=', 4),
     (code1, 'f(a,b,xy=', None),
-    (code1, 'f(u=', (0, None)),
+    (code1, 'f(u=', None),
     (code1, 'f(v=', 1),
 
     # **kwargs
@@ -438,7 +450,7 @@ _calls = [
     (code2, 'g(a,b,abc=1,abd=4,abd=', 5),
     (code2, 'g(a,b,kw', 5),
     (code2, 'g(a,b,kwargs=', 5),
-    (code2, 'g(u=', (0, 5)),
+    (code2, 'g(u=', 5),
     (code2, 'g(v=', 1),
 
     # *args
@@ -450,7 +462,7 @@ _calls = [
     (code3, 'h(a,b,c,(3,)', 2),
     (code3, 'h(a,b,args=', None),
     (code3, 'h(u,v=', 1),
-    (code3, 'h(u=', (0, None)),
+    (code3, 'h(u=', None),
     (code3, 'h(u,*xxx', 1),
     (code3, 'h(u,*xxx,*yyy', 1),
     (code3, 'h(u,*[]', 1),
@@ -483,7 +495,7 @@ _calls = [
     (code4, 'i(1, [a?b,*', 2),
     (code4, 'i(?b,*r,c', 1),
     (code4, 'i(?*', 0),
-    (code4, 'i(?**', (0, 1)),
+    (code4, 'i(?**', 1),
 
     # Random
     (code4, 'i(()', 0),
@@ -497,11 +509,6 @@ _calls = [
 @pytest.mark.parametrize('ending', ['', ')'])
 @pytest.mark.parametrize('code, call, expected_index', _calls)
 def test_signature_index(Script, environment, code, call, expected_index, ending):
-    if isinstance(expected_index, tuple):
-        expected_index = expected_index[environment.version_info > (3, 8)]
-    if environment.version_info < (3, 8):
-        code = code.replace('/,', '')
-
     sig, = Script(code + '\n' + call + ending).get_signatures(column=len(call))
     index = sig.index
     assert expected_index == index