+++ CHANGELOG.rst
@@ -6,6 +6,17 @@ Changelog
 Unreleased
 ++++++++++
 
+0.20.0 (2026-05-02)
++++++++++++++++++++
+
+- Python 3.14 support
+- Removed support for Python 3.8 and 3.9
+- Upgraded Typeshed
+- Better support for Final/ClassVar
+- ``__new__`` is now also recognized as a signature and TypeVar inference
+- Support for ``Self``
+- Support for ``TypeAlias``, generics for ``type[...]`` and ``tuple[...]``
+
 0.19.2 (2024-11-10)
 +++++++++++++++++++